requirements.txt
# Web Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart

# RAG & AI
sentence-transformers==2.6.0
faiss-cpu==1.8.0
google-generativeai==0.4.1
tiktoken==0.6.0

# Configuration & Utilities
pydantic>=2.6.0
pydantic-settings==2.2.1